requirements.txt
astroid==1.4.9
Chameleon==3.1
factory-boy==2.8.1
Faker==0.7.10
hupper==0.4.4
isort==4.2.5
Jinja2==2.9.5
lazy-object-proxy==1.2.2
MarkupSafe==1.0
mccabe==0.6.1
PasteDeploy==1.5.2
pylint==1.6.5
pyramid==1.8.3
pyramid-chameleon==0.3
pyramid-jinja2==2.7
python-dateutil==2.6.0
repoze.lru==0.6
requests==2.13.0
six==1.10.0
translationstring==1.3
venusian==1.0
vobject==0.9.4.1
WebOb==1.7.2
wrapt==1.10.10
zope.deprecation==4.2.0
zope.interface==4.3.3
